requirements.txt
speechrecognition
deep-translator
gtts
pygame
requests
google-generativeai
langdetect
langid
pyaudio
python-dotenv
googletrans